:recycle: Move student specific application serialization to its own file.
Why? To leave room for more standard REST style object serialization.
